perf: Optimize first screen loading
